exif_filter/requirements.txt
boto3
requests
Pillow
exif
retry